config/requirements.txt
streamlit==1.13.0
streamlit-authenticator==0.2.1
pyyaml==6.0